chore: moving wrappers out of analysis
